Extract shared logic into common.js
