+++ packages/cli-kit/src/api/admin.ts
@@ -1,8 +1,8 @@
-import {buildHeaders, sanitizedHeadersOutput} from './common.js'
+import {buildHeaders, debugLogRequest, handlingErrors} from './common.js'
 import {AdminSession} from '../session.js'
 import {debug, content, token as outputToken} from '../output.js'
 import {Bug, Abort} from '../error.js'
-import {request as graphqlRequest, gql, RequestDocument, Variables, ClientError} from 'graphql-request'
+import {request as graphqlRequest, gql, RequestDocument, Variables} from 'graphql-request'
 
 const UnauthorizedAccessError = (store: string) => {
   const adminLink = outputToken.link(`URL`, `https://${store}/admin`)
@@ -20,36 +20,15 @@ const UnknownError = () => {
 }
 
 export async function request<T>(query: RequestDocument, session: AdminSession, variables?: Variables): Promise<T> {
-  const version = await fetchApiVersion(session)
-  const url = adminUrl(session.storeFqdn, version)
-  const headers = await buildHeaders(session.token)
-  debug(`
-Sending Admin GraphQL request:
-${query}
-
-With variables:
-${variables ? JSON.stringify(variables, null, 2) : ''}
-
-And headers:
-${sanitizedHeadersOutput(headers)}
-`)
-  try {
+  const api = 'Admin'
+  return handlingErrors(api, async () => {
+    const version = await fetchApiVersion(session)
+    const url = adminUrl(session.storeFqdn, version)
+    const headers = await buildHeaders(session.token)
+    debugLogRequest(api, query, variables, headers)
     const response = await graphqlRequest<T>(url, query, variables, headers)
     return response
-  } catch (error) {
-    if (error instanceof ClientError) {
-      const errorMessage = content`
-The Admin GraphQL API responded unsuccessfully with the HTTP status ${`${error.response.status}`} and errors:
-
-${outputToken.json(error.response.errors)}
-      `
-      const abortError = new Abort(errorMessage.value)
-      abortError.stack = error.stack
-      throw abortError
-    } else {
-      throw error
-    }
-  }
+  })
 }
 
 async function fetchApiVersion(session: AdminSession): Promise<string> {
