+++ packages/cli-kit/src/api/partners.test.ts
@@ -13,7 +13,13 @@ vi.mock('graphql-request', async () => {
   }
 })
 
-vi.mock('./common')
+vi.mock('./common.js', async () => {
+  const common: any = await vi.importActual('./common.js')
+  return {
+    ...common,
+    buildHeaders: vi.fn(),
+  }
+})
 vi.mock('../environment/fqdn')
 
 const mockedResult = 'OK'
